# shared/contracts/openapi/batch-downloader-api.yaml
openapi: 3.1.0
info:
  title: Batch Downloader API
  version: 1.0.0
  description: API pública versionada para identidad, catálogo y trabajos de descarga.
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /admin/apps:
    servers:
      - url: /
    get:
      operationId: listAdminApps
      summary: Listar aplicaciones para la mesa de administración
      tags: [Apps admin]
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: status
          in: query
          description: unresolved se resuelve internamente como review OR missing.
          schema:
            type: string
            enum: [all, unresolved, available, review, missing]
            default: unresolved
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, updated, downloads]
            default: updated
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página administrativa de aplicaciones.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/app-discoveries:
    servers:
      - url: /
    post:
      operationId: createWebsiteAppDiscovery
      summary: Analizar una web oficial y buscar instaladores verificables
      tags: [Apps admin]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebsiteAppDiscoveryRequest'
      responses:
        '202':
          description: Descubrimiento encolado o envío idéntico reutilizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteAppDiscovery'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/app-discoveries/{discoveryId}:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/WebsiteDiscoveryId'
    get:
      operationId: getWebsiteAppDiscovery
      summary: Consultar el progreso y la previsualización de un alta automática
      tags: [Apps admin]
      responses:
        '200':
          description: Evidencia segura; nunca contiene las URL de los ejecutables.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteAppDiscovery'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/app-discoveries/{discoveryId}/apply:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/WebsiteDiscoveryId'
    post:
      operationId: applyWebsiteAppDiscovery
      summary: Revalidar instaladores y crear la aplicación desde la previsualización
      tags: [Apps admin]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebsiteAppDiscoveryApplyRequest'
      responses:
        '200':
          description: Aplicación creada; queda disponible solo con instaladores válidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteAppDiscoveryApplyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/apps/{appId}/manual-installer-inspections:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/AppId'
    post:
      operationId: createManualInstallerInspection
      summary: Encolar una inspección manual de instaladores
      tags: [Apps admin]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualInstallerInspectionRequest'
      responses:
        '202':
          description: Inspección encolada o envío idéntico reutilizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualInstallerInspection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/apps/{appId}/manual-installer-inspections/current:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/AppId'
    get:
      operationId: getCurrentManualInstallerInspection
      summary: Recuperar la inspección manual abierta más reciente
      tags: [Apps admin]
      responses:
        '200':
          description: Inspección abierta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualInstallerInspection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/apps/{appId}/manual-installer-inspections/{inspectionId}:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/AppId'
      - $ref: '#/components/parameters/InspectionId'
    get:
      operationId: getManualInstallerInspection
      summary: Consultar progreso y previsualización de una inspección manual
      tags: [Apps admin]
      responses:
        '200':
          description: Estado seguro de la inspección; nunca contiene las URL finales.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualInstallerInspection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/apps/{appId}/manual-installer-inspections/{inspectionId}/apply:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/AppId'
      - $ref: '#/components/parameters/InspectionId'
    post:
      operationId: applyManualInstallerInspection
      summary: Revalidar, guardar y publicar los instaladores de una inspección manual
      tags: [Apps admin]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualInstallerApplyRequest'
      responses:
        '200':
          description: Aplicación publicada atómicamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualInstallerApplyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /admin/apps/absence-verifications/summary:
    servers:
      - url: /
    get:
      operationId: getInstallerAbsenceVerificationSummary
      summary: Contar ausencias activas y missing sin evidencia
      tags: [Apps admin]
      responses:
        '200':
          description: Criterio de cierre autoritativo de la campaña.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallerAbsenceVerificationSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/apps/{appId}/absence-verification:
    servers:
      - url: /
    parameters:
      - $ref: '#/components/parameters/AppId'
    get:
      operationId: getActiveInstallerAbsenceVerification
      summary: Obtener la evidencia negativa activa de una aplicación
      tags: [Apps admin]
      responses:
        '200':
          description: Acta activa o null.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/InstallerAbsenceVerification'
                  - type: 'null'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: confirmInstallerAbsence
      summary: Confirmar que no existe un binario instalable tras revisar todas las fuentes
      tags: [Apps admin]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallerAbsenceVerificationRequest'
      responses:
        '201':
          description: Evidencia persistente creada y aplicación clasificada como missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallerAbsenceVerification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BATCH_SESSION
    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    AppId:
      name: appId
      in: path
      required: true
      schema:
        type: string
...
    InspectionId:
      name: inspectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WebsiteDiscoveryId:
      name: discoveryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Entrada no válida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Sesión ausente o inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: La sesión o el estado de la cuenta no permiten la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: El estado actual impide la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: La URL o el instalador no superan la validación segura.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    WebsiteAppDiscoveryRequest:
      type: object
      additionalProperties: false
      required: [officialUrl]
      properties:
        officialUrl:
          type: string
          format: uri
          pattern: '^https://'
          maxLength: 2048
          description: Web pública del producto; se cifra durante el trabajo.
        installerUrls:
          $ref: '#/components/schemas/WebsiteAppInstallerUrls'
...
    AppPage:
      type: object
      required: [data, page, pageSize, total, requestedMode, appliedMode]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/App'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        requestedMode:
          $ref: '#/components/schemas/SearchMode'
        appliedMode:
          $ref: '#/components/schemas/SearchMode'
        modelVersion:
          type: [string, 'null']
        indexVersion:
          type: [string, 'null']
        degradedReason:
          type: [string, 'null']
...